pyproject.toml
[build-system]
requires = ['hatchling', 'hatch-fancy-pypi-readme>=22.5.0', "hatch-vcs"]
build-backend = 'hatchling.build'

#######################################################################################
# Project metadata
#######################################################################################
[project]
name = "tsdb"
requires-python = ">=3.12"
dynamic = ["version"]
authors = [{ name = "collective.ai", email = "info@collectiveai.io" }]
maintainers = [
    { name = "jedzill4", email = "r@collectiveai.io" },
    { name = "lionelchamorro", email = "lio@collectiveai.io" },

]
description = "Timeseries DB"
classifiers = [
    # How mature is this project? Common values are
    #   3 - Alpha
    #   4 - Beta
    #   5 - Production/Stable
    'Development Status :: 4 - Beta',
    # 'Development Status :: 5 - Production/Stable',

    # Indicate who your project is intended for
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",

    # Pick your license as you wish (see also "license" above)
    "License :: OSI Approved :: MIT License",

    # Specify the Python versions you support here.
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

keywords = ["api", "deep-learning", "platform"]

#######################################################################################
# Dependencies
#######################################################################################
dependencies = [
    "alembic>=1.16.2",
    "more-itertools>=10.7.0",
    "pydantic==2.10.6",
    "pydantic-ai>=0.3.5",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-magic>=0.4.27",
    "sqlalchemy-timescaledb>=0.4.1",
    "duckdb>=0.10.0",
    "boto3>=1.34.0",
]


#######################################################################################
# Optional dependencies
#######################################################################################
[dependency-groups]
api = []
forecast = [
    "darts>=0.36.0",
]
dev = [
    "hatch>=1.14.1",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "seaborn>=0.13.2",
    "plotly>=6.2.0",
    "jupyter>=1.1.1",
    "diskcache>=5.6.3",
    "typer>=0.16.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest-asyncio>=0.25.0",
    "psycopg2-binary>=2.9.10",
]


#######################################################################################
# Project extra configuration
#######################################################################################

# Package
# =======
[tool.setuptools.packages.find]
include = ["tsdb"]

# Scripts
# =======
[project.scripts]

# Package data
# ============
[tool.setuptools.package-data]
"*" = ["*.yml", "*.yaml", "*.csv", "*.html"]

#######################################################################################
# Hatch configuration
#######################################################################################
[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "tsdb/version.py"


[tool.hatch.version.raw-options]
version_scheme = "only-version"
local_scheme = "no-local-version"
# git.branch = "main"


[tool.hatch.metadata]
allow-direct-references = true


#######################################################################################
# Pylint configuration
#######################################################################################
[tool.pylint.messages_control]
disable = "C0330, C0326"


#######################################################################################
# Isort configuration
#######################################################################################
[tool.isort]
profile = "black"
length_sort = true
combine_as_imports = true
force_sort_within_sections = true
lexicographical = true
group_by_package = true

#######################################################################################
# Pytest
#######################################################################################
[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    'ignore:.*:DeprecationWarning',
]
